pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "llm-security-framework"
version = "1.1.0"
description = "Enterprise-grade security testing framework for LLMs and RAG systems"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["llm", "security", "red-team", "owasp", "rag", "testing"]

# ── Core dependencies (always needed) ──
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "psutil>=5.9",
    "pandas>=2.0",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
# API model adapters
openai = ["openai>=1.12"]
anthropic = ["anthropic>=0.18"]
gemini = ["google-generativeai>=0.4"]
huggingface = ["huggingface-hub>=0.20"]
all-apis = ["openai>=1.12", "anthropic>=0.18", "google-generativeai>=0.4", "huggingface-hub>=0.20"]

# Local model support
local = ["llama-cpp-python>=0.2.50"]

# Dense vector retrieval (RAG upgrade)
dense = ["sentence-transformers>=2.3", "numpy>=1.24"]

# Document parsers
docs = ["PyPDF2>=3.0", "python-docx>=1.0"]

# GPU monitoring
gpu = ["pynvml>=11.5"]

# Full installation (everything)
full = [
    "openai>=1.12",
    "anthropic>=0.18",
    "google-generativeai>=0.4",
    "huggingface-hub>=0.20",
    "llama-cpp-python>=0.2.50",
    "sentence-transformers>=2.3",
    "numpy>=1.24",
    "PyPDF2>=3.0",
    "python-docx>=1.0",
    "pynvml>=11.5",
]

# Development tools
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
llm-security-test = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "adapters*"]

# ── Ruff (linting + formatting) ──
[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "adapters"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["adapters", "src"]

# ── Mypy ──
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = false  # TODO: enable once all modules have type hints
no_implicit_optional = true
strict_equality = true

# ── Pytest ──
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring API keys",
]
